pyproject.toml
[tool.poetry]
name = "editions"
version = "0.1.0"
description = "Utilities for managing musical score exports."
readme = "README.md"
authors = ["Bas"]
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
PyYAML = "6.0.1"
tqdm = "4.66.4"
music21 = "^9.9.1"
pydantic = "^2.9.2"

[tool.poetry.group.dev.dependencies]
black = "24.4.2"
pytest = "^8.3.2"

[tool.poetry.scripts]
extract = "src.extract_cli:main"
index = "src.index_cli:main"
schema = "src.schema_cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = ["-ra", "--strict-markers"]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
